Refactor test modules
This splits out some of the legacy structure into files with better names and a little more clarity of purpose
